Fix prealignment thumbnail creation
Fix montage offset handling - treat as interleaved with affines of
previous sections, instead of applying cumulative offset then
applying the cumulative affine.

Fix transform calculation to be based on the matches between a
section pair, instead of the transformed nodes of a section's mesh.

* [method_name]_APPROXIMATE methods now refer to calculations based
on the Mesh type.
* [method_name]_SOLVE methods now refer to calculations based on
the Matches type.

Translate prealignment thumbnail generation into the prealign
blockmatching script, so that alignment success can be easily
assessed and errors corrected before costly image rendering is
executed.

Reformat and update comments when issues are encountered.