Rework final rendering to allow for easy rescoping
Include FINISHED category of renderings, to occur after ALIGNED.

Aligned images are rendered in full, with an offset in the global
space, but are not scoped (padded) in the global space. Finished
images are scoped in the global space. This division allows
previously aligned images to be easily rescoped if needed. For
example, it's likely that a stack is aligned in stages. Until the
final portion of the stack is aligned, it won't be known how much
to pad all the aligned images to make a consistently sized stack.
By rescoping, the already aligned images can be properly placed
with a rescoped image size without the need to rerender.

Update render_aligned to include write_finished. Add FINISHED
index of -5.

Update view_stack to allow proper looping in all dimension
permutations.